requirements.txt
flask==2.3.3
PyGithub==1.59.1
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
python-dotenv==1.0.0
textblob==0.17.1
scikit-learn==1.3.2
python-dateutil==2.8.2
Werkzeug==2.3.7